kpanic: Add kernel panic routine
kpanic produces a kernel panic. When a kernel panic happens, the
system will print to the VGA console a message stating the error
code and possibly a panic message, and then will halt the
system.

kpanic should happen when an illegal state for the kernel is
reached or when a non-recoverable state is entered after something
wrong happened. It is similar to the blue screen of death, the
kernel panic on Linux or the guru meditation on Amiga.