Create a local interrupt handler table
Instead of having every interrupt be handled in the common
function, a handler table has been created. Using the set_handler
method, the function that handles an interrupt can be dynamically
changed in the global handler.

This way, for instance, the keyboard driver could set up his own
handler in the table.