Add basic keyboard and timer handlers
driver/keyboard.c: keyboard driver. At the moment there is no
functionality, excepting from taking the scancode from the
keyboard queue when a keyboard interrupt comes in.

driver/timer.c: timer driver. At the moment it adds ticks to
a tick counter that can be accessed later. This functionality
might change and be more complex in the future.