Replace the dist system with CD-ROM
The previous distribution system that packed the NativeOS Kernel
Image in a floppy disk is gone. It has been replaced by a new
distribution system that packs the Kernel Image in a CD-ROM thanks
to GRUB.

A few targets have been added to Makefile:

* make cdrom: will create the CD-ROM image of NativeOS.
* make qemu: now loads the CD-ROM image.
* make qemu-gdb: same as make qemu but waits for GDB.

The older targets that created IMG floppy drives have been removed
just like the old tools that were in the Tools folder for creating
these image files. mtools is no more a requirement and the old
GRUB-Legacy bootloader files can be removed from the repository.