Organize directory structure and build path
Instead of keeping .o files near the .c or .s files, a proper build
structure is added. Built things are located in out/ folder now.

The Makefile is instructed to keep the directory structure; it is just
the root path for the built artifacts the thing that has changed. So,
instead of

    kernel/driver/vga.c -> kernel/driver/vga.o

Object files are now placed on

    kernel/driver/vga.c -> out/kernel/driver/vga.o

But NOT on

    kernel/driver/vga.c -> out/vga.o (NO)

Also, the ELF image and the distribution disk is now placed on out/ as
well. Cleaning up the project should be now as easy as doing rm -rf out.