Organize kmain driver initialization
Driver initialization is starting to get messy.

Said that, a few changes have been made to how the drivers are
initialized on kmain. VGA console driver is not initialized anymore
since the current kernel requests an VBE mode to the bootloader.

Multiboot checksum verification has been defered until we actually need
to access the multiboot_info_t structure. There is a comment explaining
why we can do this: because even if the bootloader wasn't multiboot,
GDT, IDT, COM, keyboard and timer can be initialized safely because they
don't rely on multiboot_info_t.