Remove legacy code from main.c
Legacy code is code that was used back in the day for whatever reason,
usually just debugging to VGA to assert things were properly working,
but that was never moved or deleted after it wasn't required.

Most important change applied on this commit is that count_memory
function has moved to paging.c. count_memory is a static function that
parses the multiboot structure in order to count how many megabytes of
memory has the computer. Because this function is only used for setting
up the paging system, it has moved there.

I have also removed a few headers that were included in main.c but not
used and removed some dead code from kmain: screen is now kept black,
and instead of starting a dummy echo server on the serial port, an
infinite loop is entered.