Delete the old toolchain system
Previously there used to be a helper script for downloading precompiled
binaries for the i386-elf toolchain that you could invoke to make faster
the development.

Those scripts and the compiler versions have become outdated and I'm
replacing them with another way of getting the compiler. Therefore, I'm
removing the old toolchain system.

At the moment, the toolchain has to be manually supplied by the user.