Use Makefile flag for setting debug mode
This commit renames the DEBUG macro variable to HAVE_DEBUG. It is used
to control whether printk should emit characters to the VGA console or
to the serial console. It was introduced in commit 2df94e1 (#1).

The reason for this rename is to avoid possible conflicts by making
conditional flags always have a prefix (i.e., HAVE_DEBUG, HAVE_EFI,
HAVE_ACPI...)

Additionally, the debug mode is now set via Makefile. To turn on debug
mode, compile with `make DEBUG=1`.