Test GRUB magic number in boot.S
We should test for the magic number as soon as possible and stop the
execution if we detect the kernel has been loaded using something that
is not a Multiboot compatible bootloader (GNU GRUB or a similar
bootloader).

This commit performs that check using assembly code. This make
unnecesary passing that parameter to kmain. kmain has now a pointer to
the multiboot structure as a parameter.