Seed an initial ramdisk
The Makefile has been modified to create a tarfile using the contents of
the tools/initrd directory, which is a directory that has been created
with a set of test files:

* system/splash.bmp: a splash image of NativeOS, which hopefully at some
  point will be presented either by the kernel loader or by the kernel
  image during start up.

* user/readme.txt: a dummy text file welcoming the user to NativeOS.

When the CD-ROM is compiled, the contents of this directory are packed
in an U-STAR (.tar) file.  This file is also included as a kernel module
by the multiboot bootloader, and it's available to the kernel when the
kernel starts.