Add virtual NINE device file
This commit adds the nine device file. nine is a device file that when
read from, it returns as many times the ASCII character '9' as you
request, making it a simple enough device for testing. nine acts as a
sink, so it accepts whatever you write into it, but it will be dropped.

nine is not our alternative to the zero device.  nine could be at some
point our random number generator, although the problem with randomness
is that you can never be sure.