Add device file system root node
This commit creates the device file system handler.  The handler allows
devices to be registered.  The device file system handler includes a
VFS node that represents a directory with all the registered nodes.  It
can be used to retrieve a particular device from the device file system.

To write this commit, the definitions for the readdir and the finddir
operations had to be specified.  They are based in their UNIX
equivalents.

* readdir returns the Nth node contained in a particular directory node.
  This operation is only available in directory nodes.  An index is
  given as parameter.  By calling this function with an incremented
  index, on each iteration a file part of the file system will be
  retrieved.  When NULL is retrieved, it means that there are no more
  files in this directory.

* finddir returns the node with the given name in this directory.  This
  operation is only available in directory nodes.  A pointer to a string
  with the name of a node to search is provided.  If the directory
  contains a node with this name, a pointer to that node will be
  returned.  Otherwise, NULL will be returned.

At the moment, readdir and finddir has support for . and .. special
names. This means that:

* Calling readdir with an index == 0 will return . (the directory root
  itself). Same thing happens when calling readdir with an index == 1,
  because the directory root has no parent. Consecutive calls will
  return additional nodes.

* Calling finddir looking for a file named either "." or ".." is
  successful and returns a pointer to the directory root itself.