Add ELF header parser
elf32_header is a struct that represents an ELF object header, including
header fields such as the magic number and other meta-information about
the ELF file itself.

The elf_is_valid() function is able to parse whether the given ELF
header is valid or not. The expected use case for this function is to
parse a blob that supposedly represents an ELF object, and test whether
the ELF header is found to be valid or not, to detect rogue blobs.