Implement a basic symbol table
The symbol table is used to register public symbols present in ELF
object files as well as the memory addresses for these symbols.  Once
this feature gets finalized, the symbol table will be loaded with the
public functions available in nativeos.elf as well as other kernel
extensions.

Kernel extensions are additional and relocatable ELF objects.  When a
kernel extension is loaded, references to functions present in
nativeos.elf are not initialized because the kernel extension does not
know the memory address of the function inside nativeos.elf.

The symbol table will keep references to each public function exported
by nativeos.elf, and the relocation process will replace the dummy
memory addresses present in the ELF object by their actual values.