Parse the kernel ELF image symbol table
symtab_load_elf_symtab is a function that loads the contents of an ELF
symbol table into the kernel symbol table.  This is used to parse the
contents of the kernel object itself (nativeos.elf) symbol table in
order to have an in memory index of the public functions available in
the kernel and the memory addresses of such functions.

It is possible now to get the memory address of public functions
declared in nativeos.elf, such as kmain or heap_alloc. These functions
are used by kernel extensions, but since the kernel extensions cannot
know the exact memory address of these functions at compile time, the
kernel symbol table is required to relocate these extension files.