Add dummy kernel extension module
This kernel extension module will probably dissapear before merging this
branch into master, or be replaced with a better "test" kernel
extension.

Kernel extensions are subdirectories placed in ext/. Each subdirectory
compiles to a different .kxt binary image, representing a kernel
extension. Kernel extensions are ELF binary images that are loaded at
boot time. The NativeOS loader is expected to load this modules and to
run the start up code present on them.

At the moment, adder.kxt is a test kernel extension that provides a few
services for making integer arithmetic. The following services are
exported by adder.kxt:

* add: adds two numbers.
* sub: subtracts two numbers
* len: returns the length of a hardcoded string.

The first two services test the ability of the kernel to relocate and
execute ELF functions. The last service makes use of a shared kernel
function, and tests the ability to relocate external functions exported
by the kernel, and the ability to use multiple segments.