Fix kernel extensions being added to nativeos.elf
Although the plan is fine, the execution is wrong.  The purpose of
having $(KEXT_OBJS) as a dependency is to force make to build the kernel
extensions at some point.

However, out/nativeos.elf is not the best place for adding this
dependency.  There are a few reasons for this.

1. Building the core nativeos.elf should not depend on building the
   kernel extensions (that's the whole point of extensions!!!!).

2. Most importantly, because of the way the $(LD) call made in the
   recipe for out/nativeos.elf is written --using  $^ , $(KEXT_OBJS)
   would enter into the build command, so the kernel extensions will be
   linked into nativeos.elf as well.

Therefore, this commit removes $(KEXT_OBJS) from the out/nativeos.elf.
It has been checked that out/nativeos.elf is now kext-free. $(KEXT_OBJS)
has now been added as a dependency for nativeos.iso, as it should,
because after all, the ISO file is what actually requires these .kxt
files.