Allocate NOBITS + ALLOC section headers
This commit will process the section headers in a kernel extension
marked as NOBITS + ALLOC and allocate those section headers.

The most important use case for these kinds of sections are the .bss and
the .sbbs sections found in most ELF images that make use of global
variables and static variables that are not initialized or that are
initialized to zero.  Instead of writing a lot of zeros into the program
file, the spec strips the contents of that section leaving only the
header.  The program loader has to allocate the space for that section,
and update the section header to point to the new buffer in the offset
field.