doc: rename 8259 PICs to PIC1 and PIC2
This commit rephrases some comments that reference the two 8259 PIC
controllers present in the x86 architecture so that their names are PIC1
and PIC2.

PIC1 is attached to the CPU and it is managed through I/O Ports 0x0020
and 0x0021. PIC2 is attached to PIC1 in a cascading way, and it is
managed through I/O Ports 0x00A0 and 0x00A1.