stdkern: rebrand libc as stdkern
The existing libc implementation is not actually a libc, but rather a
subset of the C library to polyfill the functions required by the kernel
in an environment where it is not feasible to have a standard library.

Although it was eventually considered that the libc folder could be used
to build polyfills for the kernel image and for some userland
applications, it is better to drop the idea and target a different libc
for userland applications.

Therefore, this commit will inline the libc into a kernel component
called stdkern. The stdkern contains standard functions used by the
kernel, such as the C standard library subset required by the kernel to
do its work.

The header files have been merged into <sys/stdkern.h>. The name has
been intentionally picked to avoid confusion when including the header
file in other C source code files: this is not a standard library, but
rather a subset required by the kernel to do its job.

Because this commit moves and rewrites files, a well-tested clang-format
that has been used in other projects in the past is added in this commit
too, hoping to fix for once having multiple source code format
conventions.