i386: inline GDT initialization
Although it could be a cool experiment to play with the capabilities of
the GDT (hey, I know that segmented modes haven't been trendy in 30
years, but they might have some point!), this commit will inline the
load of the GDT table by removing all the functions and structs that
would fill a table in C source codes and replace it with inline assembly
initialization.

The contents of the table were gathered by compiling a kernel image,
loading it with a debugger and inspecting the contents of the GDT table
right after the gdt_load function was executed.

Since the GDT is now set up before kmain is called, it makes sense to
remove gdt.c, gdt.h and lgdt.S since they are not required anymore.