i386: force -m32 flag on GCC
This commit will make possible to use the host GCC compiler on ELF based
environments such as GNU/Linux on the amd64 platform.

It was previously possible to use the host GCC compiler on GNU/Linux on
Linux-i386, but on Linux-amd64 it fails because host GCC will try to
compile a 64 bit kernel, even when using the I386 profile.