i386: switch from GCC to LLVM to build the kernel
This commit will replace GCC as the compiler to build the system with
clang from the LLVM compiler suite.

The reason I did this is because due to the way LLVM is designed, it is
possible to build code for foreign ABIs such as i386-elf on an x86_64
macOS without requiring a cross-compiler, making it easier to bootstrap
a development environment. Having support for the goodies that clang
includes such as clangd and clang-format is nice. GCC is a great
compiler, but it lags on these improvements.

NativeOS compiles out of the box already with clang. However, it doesn't
link well. It seems that clang will try to add a section at the
beginning of the ELF file called .note.gnu.build-id. This causes the
.text section to fall outside of the 4 KB region that Multiboot
bootloaders use to find for the magic header. Therefore, this commit:

* Will patch the ldscript so that the .note.gnu.build-id is added after
  the .text section. I don't want to remove this section either because
  it seems that it can be used to do cool stuff, and this is something
  that I want to explore in the future.

* Will move the multiboot header into its own section called .multiboot.
  This section will be placed before the .text section, to separate the
  header from the code.