stdkern: add doubly linked list
A simple double linked list is based using the standard textbook
implementation of linked lists. It is built on top of the memory heap
and uses the heap allocation functions.

A lot of functions for adding, retrieving and deleting elements from a
linked list is added. The functions are listed in the sys/list.h file.
They have been tested in a test suite that this commit will not add
because I want to set up a proper testing platform before committing the
test cases.