stdkern: add a ring buffer
A ring buffer is a circular FIFO structure that holds a byte buffer.

By calling the write function, it is possible to save data in the byte
buffer. The data can be read using the read function. Because this is a
FIFO data structure, reading data will also remove it from the data
structure; thus, two consecutive calls to the read function may yield
two different bytes depending on the values previously written into the
buffer using the write function.

The ring buffer has an overflow logic to detect whether a previous write
operation did overflow the buffer.

Note that this is a first iteration of the ring buffer and it may be
enhanced in the future as code that makes use of the ring buffer is
added to the kernel.