devmgr: draft a device manager
A device is a specific piece of hardware exposed through a standard
interface, allowing the kernel to see hardware as a set of functions,
regardless of their internal kind.

This commit will add a device manager. The device manager is internally
a list with functions to add, remove and fetch devices. When installing
a device, the device is bound to a specific name, which is used later to
refer to the specific device.

At the moment, a device has a set of functions and callbacks used by the
kernel to interact with the hardware device:

* read: allows the kernel to read data from the device.
* write: allows the kernel to send data to the device.

For example, a piece of hardware such as a serial port may be controlled
using a device. The read and write functions in the device may map to
internal functions used by the driver, which will exchange a buffer of
data through the pins of the serial port.

A terminal may be represented as a device as well. Calling the read
function on the terminal device may cause the recently pressed keys in
the keybaord to be fetched so that the kernel can see what has been
typed. Calling the write function on the terminal device could cause
some characters to be sent to the video memory in order to make them
visible in the screen.

There is a distinction between a device and a driver. The driver is the
underlying kernel module that manages pieces of hardware of a specific
family. Some driver examples:

* COM driver: may manage serial ports.
* PS2KBD driver: may manage standard PS2 keybaords.
* CGACON driver: may manage standard CGA text consoles.

When initialized, a driver may spawn zero, one or multiple devices:

* The CGACON driver may spawn a device called FB to represent the
  framebuffer. Writing to this file will send characters to the screen.
  There is only one device in this driver, since there is usually only a
  single framebuffer in the video memory.

* The COM driver may spawn multiple devices called COM1, COM2, COM3...
  to represent the different COM ports on a machine, one per serial port
  (assuming the computer has multiple serial ports available).
  Calling read or write in a specific device may cause data to be sent
  into a specific serial port.