vgatext: add the VGA driver as a devmgr module
This commit adds a text mode VGA driver exposed as a driver module. The
driver will install a device called "fb". It is possible to write text
onto the "fb" device, and it will be sent to the screen.

The source code of the module, present in vgatext.c, is based on the old
vga.c source code, and has been tweaked a little to fit into the new
APIs.