fs: draft a virtual file system
Instead of implementing functions that are linked to a specific file
system driver, most operating systems will prefer to implement a virtual
file system, which is an abstraction layer that dehumanizes each file
system implementation into a set of common structs with a set of
function pointers that are defined by each file system implementation.

This commit adds a virtual file system basic initialization code.
Everything lies into a vfs_node, which is a data structure that
represents a node. It has some common metadata: name, class... and a few
function pointers which are set by each file system implementation.

A node has a class:

* It might be a file. The pointers can be used to get the bytes stored
  in the file, or to append or replace data in the file contents. This
  byte container can be backed by a persistent layer such as a hard
  disk.

* It might be a device. This is an ancestral hack used by many operating
  systems to expose hardware using the file system. The read and write
  function pointers are handled by the hardware driver itself, to
  translate read/write operations at the FS layer into read/write
  operations at the hardware IO layer.

* It might be a directory. A directory contains more nodes in a
  hierarchy fashion. The function pointers can be used to traverse the
  directory and get the child nodes, in a recursive data structure.

Unlike most UNIX operating systems, NativeOS will not use a hierarchical
file system structure. Instead, each file system driver can mount a
volume, which is a glorified root directory. Each volume has a separate
root directory node, identified by some mountname, such as DEV for the
devices volume or TMP for the temporal file system volume.

struct mountpoint is an internal structure added to kern/fs_vfs.c which
holds the metadata for a volume, including the root and the mountname.
The remains of kern/fs_vfs.c is used to manage a linked list that allows
mounting and unmounting volumes from the kernel, and traversing the
list.