devmgr: add open/close hooks to device_t
Prior to combining the devmgr and the devfs, it is important to add the
newly implemented open/close file system function pointers to the
devices.

Some devices won't need an open/close system, but some devices might
want to check the flags used when calling the vfs_open function - for
instance, to prevent a read only device from being opened in write mode,
or to ask for exclusive or non-blocking mode.