stdkern: add strsep standard library function
strsep(3) is a recent function part of the C standard library that will
split a string in two halfs based on the list of token separators to
match. For instance, given the following code,

    char *input = strdup("hello.world");
    char *ret = strsep(&input, ".");

strsep(3) will attempt to split the input string using the dot as a
delimiter. When there is a match, strsep returns the first half of the
match, and updates the value of the first argument (a double pointer) so
that it points now to the second half of the string. So in the previous
example, input should be "world", and ret should be "hello".

Note that the string pointed by the first parameter must not be
read-only, but allocated in the heap or somewhere writable, because
strsep will modify the string to place a \0 at the location of the first
match.

There can be multiple separators and it will find the first ocurrence.
If no matches happen, it treats one of the halfs as NULL, and the other
half is the entire string.

The following patch can be used to write a main function that unit tests
this function in order to assert that it behaves as expected. When
compiled using `cc kernel/stdkern/strsep.c -Ikernel -DDEBUG`, the
execution of the generated program will print multiple lines, and each
of the printed lines must be true.

    diff --git a/kernel/stdkern/strsep.c b/kernel/stdkern/strsep.c
    index d136b2d..024b903 100644
    --- a/kernel/stdkern/strsep.c
    +++ b/kernel/stdkern/strsep.c
    @@ -46,3 +46,58 @@ strsep(char **strptr, const char *delimiter)
     	*strptr = 0;
     	return origstrptr;
     }
    +
    +#ifdef DEBUG
    +#include <stdio.h>
    +#include <string.h>
    +
    +int
    +main(int argc, char **argv)
    +{
    +	char *string, *retval;
    +
    +	// Normal use case: should split the string
    +	string = strdup("simple;string");
    +	retval = strsep(&string, ";");
    +	printf("iop: %s == string\n", string);
    +	printf("ret: %s == simple\n", retval);
    +
    +	// Character to locate not in delimiter list
    +	string = strdup("simple;string");
    +	retval = strsep(&string, "-");
    +	printf("iot: %lu == 0\n", (uintptr_t) string);
    +	printf("ret: %s == simple;string\n", retval);
    +
    +	// Empty delimiter list
    +	string = strdup("simple;string");
    +	retval = strsep(&string, "");
    +	printf("iot: %lu == 0\n", (uintptr_t) string);
    +	printf("ret: %s == simple;string\n", retval);
    +
    +	// Empty string is just like another string
    +	string = strdup("");
    +	retval = strsep(&string, ";");
    +	printf("iot: %lu == 0\n", (uintptr_t) string);
    +	printf("ret: '%s' == '' \n", retval);
    +
    +	// Empty string and haystack
    +	string = strdup("");
    +	retval = strsep(&string, "");
    +	printf("iot: %lu == 0\n", (uintptr_t) string);
    +	printf("ret: '%s' == '' \n", retval);
    +
    +	// Null string, returns null everywhere
    +	string = 0;
    +	retval = strsep(&string, "");
    +	printf("iot: %lu == 0\n", (uintptr_t) string);
    +	printf("ret: %lu == 0\n", (uintptr_t) retval);
    +
    +	// Combined example
    +	string = strdup("hello good morning");
    +	printf("combined: '");
    +	while ((retval = strsep(&string, " ")) != 0) {
    +		printf("%s ", retval);
    +	}
    +	printf("\b' == 'hello good morning'\n");
    +}
    +#endif