fs: add vfs_open_path function
vfs_open_path is a critical function that will lookup for a specific
node in any of the mounted file systems by traversing the volume list,
given the path for a file we want to lookup. If the path does not point
to a valid file or if the path is not valid, it should return NULL.

I don't expect this function to be error-free. I expect to find subtle
bugs and to see this function explode in surprising ways in the future.
It should be tested more carefully eventually, but it works right now.