vfs: fs_* functions family
The fs_* functions provide an abstraction layer to make calling VFS
related functions easier, and also makes the checks to assert that the
requested node properly supports the given operation.

Each function maps to an existing VFS vn_ operation, and delegates the
result and side effects to those implemented in the operation. However,
the fs_ functions family will also return a default invalid value if the
given node does not support the requested operation, such as trying to
readdir from a regular file.