pctimer: backport driver from stash
pctimer is the driver for the Intel 8253 PIT, a timer historically used
in the PC platform. This timer has multiple profiles and has an internal
clock that generates timed signals that the computer can decode to
understand the pass of time. One of the most important uses of this
device is to generate a periodic signal an amount of times per second,
that can be used to do multitasking.

This commit will backport and remove the driver previously located in
the stash, so that it is available in the kernel again. The driver has
gained support for a character device exposed at DEV:/pit, which will
allow readers to see the amount of ticks, which is meant as a debug
tool.

The benefit of backporting this driver already is that I will be able
now to put a breakpoint in the fallback_handler function without having
the breakpoint being triggered by the timer a few times per second.