vfs: rename vfs_open_path -> fs_resolve
Incorrectly calling vfs_open_path set the expectation that this function
would also open the VFS node before returning it, but it is not the
case. The function has been renamed as fs_resolve, which indicates that
this function will only return the proper VFS node after decoding the
given path, but it is up to the client to call fs_open in the node.

To make things easier, I've added to kern_main.c a function called
fs_resolve_and_open, which is going to be helpful as long as I keep
manually doing dirty things with the VFS in the kern_main function, but
this function will be replaced by a proper service layer in the future.