uart8250: backport enhanced serial driver
UART8250 is the most basic UART (serial port) that can be present in the
PC platform. It was replaced by the 16550 even before the nineties, yet
here I am using the older version just for the sake of "simplicity" and
"compatibility".

Why would I backport again the support for a device so old that cannot
even be used in real computers nowadays due to the lack of RS232
interfaces on most computers?

Most importantly, debug purposes. Any emulator can forward the COM port
to the standard output and standard input, and therefore it is a very
useful port for throwing information that should be presented in a way
that can be seen by the host.

Also, adding a new device means adding a new character device. It will
be useful to play with another device in the future, for instance, to
make sure that it is possible to use multiple devices at the same time.

This commit removes the old COM driver present in contrib/stash/drivers.
The new driver has enhancements that were not present in the old COM
driver. For instance, the newer driver uses interrupts, FIFOS and
ringbuffers, so it is possible to asynchronouslly fill a ringbuffer
until the driver properly calls the .read function to read the contents
of the buffer.