i386: add qemu --kernel to the i386 Makefile
By adding extra rules to the Makefile.i386 files, it will be possible
now to directly spawn QEMU using the -kernel flag without needing to
build a CD-ROM. This is going to make debugging the kernel faster.

QEMU kernel support is very limited. For instance, it does not support
Multiboot 2 at the moment. The multiboot structure dumped in the EBX
register is unreliable. However, depending on what I am looking to test,
it can be successful anyway and it will be faster than building an
entire CD-ROM.