vgatext: fix cursor not appearing
Not a big deal but I finally discovered why the cursor does not appear
in the vgatext driver. It turns out that the cursor uses the same
foreground color as the attribute for the character the cursor is in. So
if the attribute is set to 0x00, the cursor will not be visible because
it is black over black.

This commit will fix the behaviour by initialising the text framebuffer
to empty white over black characters when vgatext_init() is called,
making the cursor automatically appear as expected.