vfs: add VO_FBINARY flag
VO_FBINARY allows to open a file in binary mode. Some operating systems
and device drivers use this flag as a hint that affects reading and
writing from a file. Most notably, on MS-Windows-NT, reading or writing
a \n makes it handle as if it was a \r\n unless binary mode is set.

As a testbed, the uart8250 driver has been modified to understand this
hint. If the UART port is opened in non-binary mode, some translations
will be made. Most notably, sending a CR character will cause a CRLF
byte to be sent instead, and sending a DEL character will map to a BS
character.