devmgr: dynamically init drivers
This commit defines a macro called DEVICE_DESCRIPTOR that is used to
export a driver descriptor. The macro accepts two parameters: the device
name (used mainly to export the symbol) and the driver_t variable.
Internally, this is expanded into a global device_t* variable that
points to the given device_t structure.

In the macro definition there is a compiler attribute that will collect
these global variables into a separate ELF section called .text.driver.
The .text.driver section is full of pointers to these device_t driver
definitions.

When the device manager is starting, it will traverse the pointer list
and it will call dv_init() for every driver_t* structure that it finds.
Essentially this initialises all the drivers that are compiled and
linked when building the kernel.

I initially wanted to use an implementation that doesn't rely on
centinel markers in the ldscript (devices_start + devices_end).
Unfortunately, I have got used to test and debug my kernel using QEMU's
native -kernel flag, and QEMU's multiboot implementation does not expose
ELF information through the header, so I cannot get the .text.driver
section by inspecting the ELF file.