vfs: add offset to vfs_write and vfs_read calls
Calling vfs_write and vfs_read will now require to provide the offset
to use when writing or reading data. For character devices, this is
expected to be 0, since these devices are streams of bytes.

This is a breaking change, and the commit fixes the cases where the
function prototype must be updated.