devmgr: add support for block devices
This is another commit that will make changes to a lot of files at the
same time to fix some broken behaviour after introducing a code change.

The purpose of this commit is to add support for block devices. A block
device is a device that acts as a buffer of bytes, rather than as a
stream of bytes (such as a character device). At the moment, block
devices are using the standard VFS read/write prototypes, so the client
code will specify offsets when calling the device read or device write
functions.

chardev_t has been renamed to device_t. A device is either a block
device or a character device. Note that this has to do with the VFS
layer and not with the driver layer.

Taking the time to refactor the code, the following identifiers have
changed names in the fields of the related structs:

* driver_t fields now start with drv_ rather than with dv_.
* chardev_t/device_t fields now start with dev_ rather than with cd_.

Then, cd_write is now known as dev_write_chr, and it is implemented only
by character devices. Similarly, cd_read is now known as dev_read_chr
and it should only be implemented by character devices. These work as
usual and do not accept offsets when calling the functions.

The new fields for block devices are dev_write_blk and dev_read_blk,
which should be implemented by devices that behave as a block device. As
said, these functions use the same prototype than the standard VFS
prototypes, so the offset has to be provided when calling these
primitives.

The DEVFS file system driver has been modified so that calling the
read/write function of the DEVFS driver will forward the call to the
appropiate device_t function depending on whether the device family is a
block device or a character device.