vtcon: create write-only console driver
The console is a special device that manages access to an input stream
(such as the keyboard or the mouse) and to an output stream (such as the
framebuffer), allowing an operator to type characters using the keyboard
and have them picked by the system, or sending characters to the
framebuffer.

This driver is still on its early development and a lot of things are
missing. Some features that have to be implemented in the future are:

* Support for some limited ANSI escape codes, so that the output can be
  properly formatted.
* Add ioctl at the VFS layer so that I can move some VGA hardware
  operations that are part of vtcon at the moment to the vgafb driver.
* Support for reading from the keyboard device and decoding the
  scancodes into something more useful.