vgafb: add extra operations through ioctl
Some operations that are part of the VGA, such as setting or unsetting
the cursor visibility or moving the cursor around the screen, are now
implemented using ioctl operations, instead of using proxy methods in
the vtcon controller. The vtcon controller will now send the requests as
ioctl requests instead.

The implemented list of I/O requests for the vgafb device are:

* VGAFB_IOCTL_SETCUR (0): used to toggle the visibility of the cursor
  depending on the given argument.
* VGAFB_IOCTL_MOVECUR (1): used to move the cursor to an absolute offset
  of the screen, relative to the top-level.