vtcon: decode keyboard scancodes
This commit adds keyboard reading capabilities to the vtcon driver.
Reading from the vtcon device will transfer the request to the keyboard.
At the moment, the VK device will generate keyboard events as a result
of calling, so the protocol is not text, but binary.

However, these keyboard events are already decoded and they contain
information about the keyboard state (such as whether the SHIFT, ALT or
CTRL keys are down or up), and it will perform some basic keyboard
layout transformations to convert a scancode into the proper key or
virtual key, taking into account the state of the keyboard (for
instance, if "Q" is pressed, it will pick "q" or "Q" based in the state
of the SHIFT key).

Additionally, the read call will echo to the screen the read character
as a side effect. This might change in the future, as more code is built
on top of the vtcon driver.