i386: extend Makefile to include compile dir
The compile dir will be included in the standard inclusion path for the
C compiler when building files. This will allow in the future to have a
"machine" symlinked directory, something that was added priorly in
commit ae13ca2 ("kcons: create machine/ include dir").

Also, I am considering to eventually build a dependency database for
header files, because modifying a header file doesn't currently rebuild
the source files that include those header files, so this commit may
allow to import other special things in the future.