i386: additional port I/O functions
This commit adds additional C functions that translate into `iow` and
`iol` assembly instructions. At the moment, only a function for reading
and writing bytes into ports existed, but now there is a function for
writing words (16 bit) and longs (32 bits).

io.S is one of the oldest files in the codebase, and the function names
IO_InP and IO_OutP existed since the beginning of the project. This
commit will also adapt the function names to the new code conventions
added a few years ago, so the function names are in lowercase.

Code that calls these old function names has been refactored so that it
uses the new names now, which are:

* port_out_byte and port_in_byte for the 8 bit variant.
* port_out_word and port_in_word for the 16 bit variant.
* port_out_long and port_in_long for the 32 bit variant.

This refactor moves the I/O Port code to the kernel/i386 directory as
well, so it uses the changes done previously for kcons to put the
/kernel/i386/include directory in the compile path as `machine/`.